feat(PhyTon): add display to x, y values

+++ kevin.py
@@ -4,7 +4,7 @@ class Kevin:
     def __init__(self, x, y):
         self.x = x
         self.y = y
-        self.pos = np.array([x, y])
+        self.pos = np.array([x, y], dtype=float)
         self.v = np.array([0., 0.])
         self.a = np.array([0., 0.])
